Bug Fixes
1. When checking for Clusters folder, if folder found is real folder,
not symlinked folder, it will error and ask then to delete themselves.
2. Changed the validation check for Ark App to use consumerid, not
creator id.

+++ ARK Server Manager/ServerSettingsControl.xaml.cs
@@ -680,10 +680,17 @@ namespace ARK_Server_Manager
             var profileClustersFolder = Path.Combine(Settings.InstallDirectory, Config.Default.SavedRelativePath, Config.Default.ClustersDir);
             if (Directory.Exists(profileClustersFolder))
             {
-                if (MessageBox.Show("Cluster folder already exists for this server, do you want to recreate?", "Cluster Folder Exists", MessageBoxButton.YesNo, MessageBoxImage.Question) != MessageBoxResult.Yes)
+                if (MachineUtils.IsDirectorySymbolic(profileClustersFolder))
+                {
+                    if (MessageBox.Show("Cluster folder already exists for this server, do you want to recreate?", "Cluster Folder Exists", MessageBoxButton.YesNo, MessageBoxImage.Question) != MessageBoxResult.Yes)
+                        return;
+                    Directory.Delete(profileClustersFolder);
+                }
+                else
+                {
+                    MessageBox.Show("The SymLink could not be created, a folder called Clusters already exists for this profile. Delete the Clusters folder and try again.", "Create Symlink Error", MessageBoxButton.OK, MessageBoxImage.Error);
                     return;
-
-                Directory.Delete(profileClustersFolder);
+                }
             }
 
             try
